Creacion para la pantalla de mi carrito

+++ app/src/main/java/com/devdroid07/storeapp/store/data/mappers/ProductMapper.kt
@@ -7,7 +7,7 @@ fun ProductDto.toProduct(): Product{
     return Product(
         id = id,
         title = title,
-        price = price,
+        price = price.toDouble(),
         description = description,
         category = category,
         image = image,
